feat #19 rename package to design-system
