+++ presentation/src/main/java/com/dcs/presentation/core/designsystem/CircularProgressBar.kt
@@ -1,4 +1,4 @@
-package com.dcs.presentation.core.common
+package com.dcs.presentation.core.designsystem
 
 import androidx.compose.foundation.Canvas
 import androidx.compose.foundation.layout.Box
